requirements.txt
grpcio>=1.73.1
grpcio-tools>=1.73.1
python-jose
watchdog